+++ TP_1/arithmetic.h
@@ -2,4 +2,4 @@ float SumaDosNumeros(float firstOperand,float secondOperand);
 float RestaDosNumeros(float firstOperand,float secondOperand);
 float MultiplicacionDosNumeros(float firstOperand,float secondOperand);
 float DivideDosNumeros(float firstOperand,float secondOperand);
-unsigned long long int CalculaFactorial(float number);
+unsigned long int CalculaFactorial(float number);
